+++ apps/api/src/domain/models/context/commands/add-audio-for-note/audio-added-for-note.event-handler.spec.ts
@@ -5,12 +5,15 @@ import { Test } from '@nestjs/testing';
 import buildMockConfigService from '../../../../../app/config/__tests__/utilities/buildMockConfigService';
 import buildConfigFilePath from '../../../../../app/config/buildConfigFilePath';
 import { Environment } from '../../../../../app/config/constants/environment';
+import { buildMultilingualTextFromBilingualText } from '../../../../../domain/common/build-multilingual-text-from-bilingual-text';
 import { ArangoConnectionProvider } from '../../../../../persistence/database/arango-connection.provider';
 import { ArangoDatabaseProvider } from '../../../../../persistence/database/database.provider';
 import { PersistenceModule } from '../../../../../persistence/persistence.module';
 import generateDatabaseNameForTestSuite from '../../../../../persistence/repositories/__tests__/generateDatabaseNameForTestSuite';
 import { buildTestInstance } from '../../../../../test-data/utilities';
 import buildDummyUuid from '../../../__tests__/utilities/buildDummyUuid';
+import { EventSourcedAudioItemViewModel } from '../../../audio-visual/audio-item/queries';
+import { ArangoAudioItemQueryRepository } from '../../../audio-visual/audio-item/repositories/arango-audio-item-query-repository';
 import { MultilingualAudio } from '../../../shared/multilingual-audio/multilingual-audio.entity';
 import { EventSourcedNoteViewModel } from '../../event-sourced-note.view-model';
 import { ArangoNoteQueryRepository } from '../../repositories/arango-note-query-repository';
@@ -18,29 +21,39 @@ import { INoteQueryRepository } from '../../repositories/note-query-repository.i
 import { AudioAddedForNote } from './audio-added-for-note.event';
 import { AudioAddedForNoteEventHandler } from './audio-added-for-note.event-handler';
 
+const generalContext = {
+    type: EdgeConnectionContextType.general,
+};
+
+const originalLanguageCode = LanguageCode.English;
+
+const translationLanguageCode = LanguageCode.Chilcotin;
+
 const noteId = buildDummyUuid(32);
 
-const languageCodeForAudio = LanguageCode.English;
+const audioItemIdToAdd = buildDummyUuid(4);
 
-const audioItemId = buildDummyUuid(4);
+const audioItemToAdd = buildTestInstance(EventSourcedAudioItemViewModel, {
+    id: audioItemIdToAdd,
+});
 
 const existingNoteView = buildTestInstance(EventSourcedNoteViewModel, {
     id: noteId,
     audio: MultilingualAudio.buildEmpty(),
 });
 
-const audioAddedForNote = buildTestInstance(AudioAddedForNote, {
-    payload: {
-        aggregateCompositeIdentifier: { id: noteId },
-        audioItemId,
-        languageCode: languageCodeForAudio,
-    },
+const existingAudioItemId = buildDummyUuid(123);
+
+const existingAudio = buildTestInstance(EventSourcedAudioItemViewModel, {
+    id: existingAudioItemId,
 });
 
 describe(`AudioAddedForNoteEventHandler`, () => {
     let testQueryRepository: INoteQueryRepository;
 
     let databaseProvider: ArangoDatabaseProvider;
+
+    let connectionProvider: ArangoConnectionProvider;
 
     let app: INestApplication;
 
@@ -65,7 +78,7 @@ describe(`AudioAddedForNoteEventHandler`, () => {
 
         app = moduleRef.createNestApplication();
 
-        const connectionProvider = app.get(ArangoConnectionProvider);
+        connectionProvider = app.get(ArangoConnectionProvider);
 
         databaseProvider = new ArangoDatabaseProvider(connectionProvider);
 
@@ -81,14 +94,29 @@ describe(`AudioAddedForNoteEventHandler`, () => {
     beforeEach(async () => {
         await databaseProvider.clearViews();
 
-        await testQueryRepository.createNoteAbout(
-            existingNoteView,
-            { type: ResourceType.term, id: buildDummyUuid(135) },
-            { type: EdgeConnectionContextType.general }
-        );
+        await new ArangoAudioItemQueryRepository(connectionProvider).createMany([
+            existingAudio,
+            audioItemToAdd,
+        ]);
     });
 
-    describe(`when there is a note with no audio`, () => {
+    describe(`when there is a self-note with no audio`, () => {
+        const audioAddedForNote = buildTestInstance(AudioAddedForNote, {
+            payload: {
+                aggregateCompositeIdentifier: { id: noteId },
+                audioItemId: audioItemIdToAdd,
+                languageCode: originalLanguageCode,
+            },
+        });
+
+        beforeEach(async () => {
+            await testQueryRepository.createNoteAbout(
+                existingNoteView,
+                { type: ResourceType.term, id: buildDummyUuid(135) },
+                { type: EdgeConnectionContextType.general }
+            );
+        });
+
         it(`should add audio to the note`, async () => {
             await audioAddedForNoteEventHandler.handle(audioAddedForNote);
 
@@ -96,9 +124,71 @@ describe(`AudioAddedForNoteEventHandler`, () => {
                 noteId
             )) as EventSourcedNoteViewModel;
 
-            expect(updatedView.audio.hasAudioIn(languageCodeForAudio));
+            expect(updatedView.audio.hasAudioIn(originalLanguageCode));
 
-            expect(updatedView.audio.hasAudioItem(audioItemId));
+            expect(updatedView.audio.hasAudioItem(audioItemIdToAdd));
+        });
+    });
+
+    describe(`when there is a connecting note with existing original audio`, () => {
+        const connectingNote = buildTestInstance(EventSourcedNoteViewModel, {
+            id: noteId,
+            text: buildMultilingualTextFromBilingualText(
+                {
+                    text: 'original text',
+                    languageCode: originalLanguageCode,
+                },
+                {
+                    text: 'translation into language',
+                    languageCode: translationLanguageCode,
+                }
+            ),
+
+            audio: MultilingualAudio.buildEmpty(),
+        });
+
+        const audioAddedForNote = buildTestInstance(AudioAddedForNote, {
+            payload: {
+                aggregateCompositeIdentifier: { id: noteId },
+                audioItemId: audioItemIdToAdd,
+                languageCode: translationLanguageCode,
+            },
+        });
+
+        beforeEach(async () => {
+            await testQueryRepository.connectResourcesWithNote(
+                connectingNote,
+                { type: ResourceType.term, id: buildDummyUuid(135) },
+                generalContext,
+                { type: ResourceType.song, id: buildDummyUuid(199) },
+                generalContext
+            );
+        });
+
+        it(`should add audio to the note and keep the existing audio`, async () => {
+            await testQueryRepository.addAudio(
+                connectingNote.id,
+                existingAudioItemId,
+                originalLanguageCode
+            );
+
+            await audioAddedForNoteEventHandler.handle(audioAddedForNote);
+
+            const updatedView = (await testQueryRepository.fetchById(
+                noteId
+            )) as EventSourcedNoteViewModel;
+
+            expect(updatedView.audio.hasAudioIn(originalLanguageCode)).toBe(true);
+
+            expect(updatedView.audio.getIdForAudioIn(originalLanguageCode)).toBe(
+                existingAudioItemId
+            );
+
+            expect(updatedView.audio.hasAudioIn(translationLanguageCode)).toBe(true);
+
+            expect(updatedView.audio.getIdForAudioIn(translationLanguageCode)).toBe(
+                audioItemIdToAdd
+            );
         });
     });
 });